carrefour/carrefour_1_month/requirements.txt
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
Jinja2==2.11.2
XlsxWriter==3.0.3
whitenoise==6.2.0
mysqlclient==2.1.0